Create arpitbatra1
I Created this program  it creates an address book application with features to add and delete contacts. It uses various Swing components such as JFrame, JList, JTextField, and JButton to build the graphical user interface. The Contact class is used to store contact information, and the application maintains a list of contacts.